KA - UNTESTED: Debug mode framework
Added DebugMode command, incomplete. When the robot is in Test mode, the
trigger on the left stick will control whether or not the motor(s) are
on, the slider on the left stick will control the motor speed/direction,
and the 3rd slider on the Dashboard will control which motor is being
controlled. Also, used switch-cases. Bleh.