TESTED - KA: Port numbers, fillerSolenoid
Port numbers for solenoids fixed, might need to be changed once more
Added a "fillerSolenoid" to the GearManipulator subsystem--this is
needed because Controls team is missing a part to hold in airflow. The
filler solenoid actuates along with the holderSolenoid. It should be
removed once Controls team receives the missing part. Also added buttons
to the left stick to control the pneumatics and climber in the absence
of the secondary controller, old controller buttons are commented out.