TESTED - KA: Gear override added
The main driver can now override the secondary by pressing both triggers
to drop the holder. In order to do this, I needed to get around some
(disgusting) limitations to whileHeld by turning all gear manipulator
commands into 1 default command that actuates both the holder and the
cover. The override command then literally interrupts that default
command to assume control of the holder. It controls the roof the same
way as the default command.